% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getclusterranges.R
\name{getclusterranges}
\alias{getclusterranges}
\title{Get the Cluster Ranges in a Vector of 1D Coordinates}
\usage{
getclusterranges(x, gap)
}
\arguments{
\item{x}{a numeric vector}

\item{gap}{the size for the recognation of data free gaps}
}
\description{
Get the Cluster Ranges in a Vector of 1D Coordinates
}
\examples{
x = c(1:3, 11:13)
getclusterranges(x, 3)
}
\keyword{distribution}
