% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigmmplot.R
\name{multigmmplot}
\alias{multigmmplot}
\title{Plot the Fit Results of aggregate 2-Component Gaussian Mixture Model}
\usage{
multigmmplot(x, fit_res, nbins = 15)
}
\arguments{
\item{x}{a list of a numeric vector}

\item{fit_res}{the result of AGMM}

\item{nbins}{the number of bins per cluster}
}
\description{
Plot the Fit Results of aggregate 2-Component Gaussian Mixture Model
}
\examples{
params=list(
 c(mu1=0, mu2=10, sd = 1)
 , c(mu1=10, mu2=20, sd = 1)
 )

set.seed(0)
x=lapply(
  params
  , function(v) {
    c(
      rnorm(100, mean=v[['mu1']], sd = v[['sd']])
      , rnorm(50, mean=v[['mu2']], sd = v[['sd']])
      )
  }
  )
multigmmplot(x, multigmmsamedistribu(x))
}
\keyword{distribution}
