% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigmmsamedistribulik.R
\name{multigmmsamedistribulik}
\alias{multigmmsamedistribulik}
\title{The sum of Log-Likelihoods of 1D Multi Same Distribution Gaussian Mixture Model}
\usage{
multigmmsamedistribulik(x)
}
\arguments{
\item{x}{a list of numeric vectors}
}
\description{
The sum of Log-Likelihoods of 1D Multi Same Distribution Gaussian Mixture Model
}
\examples{
set.seed(0)
x1=c(
 rnorm(100, mean=0)
 , rnorm(100, mean=1)
 )
x=list(x1)
multigmmsamedistribulik(x)(c(0.5, 1, 0.5, 1))
}
\keyword{distribution}
