% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmoutlier1d.R
\name{rmoutlier1d}
\alias{rmoutlier1d}
\title{Remove the Outliers in a Vector of 1D Coordinates}
\usage{
rmoutlier1d(x, dy_thr = dnorm(4), clustersize_thr = 3, gapsize = 10)
}
\arguments{
\item{x}{a numeric vector}

\item{dy_thr}{the threshold for dy}

\item{clustersize_thr}{the threshold for cluster size}

\item{gapsize}{the threshold of points in recognizing data free gap}
}
\description{
Remove the Outliers in a Vector of 1D Coordinates
}
\examples{
x=c(1,10:30,50)
par(mfrow=c(2,1))
plot(density(x))
plot(density(rmoutlier1d(x)))
}
\keyword{distribution}
