% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitgrp.R
\name{splitgrp}
\alias{splitgrp}
\title{Split a list with size n into groups with at least m elements}
\usage{
splitgrp(n, m)
}
\arguments{
\item{n}{an integer indicating the total length}

\item{m}{the min group size}
}
\description{
Split a list with size n into groups with at least m elements
}
\examples{

splitgrp(1, 2)
splitgrp(2, 2)
splitgrp(3, 2)
}
\keyword{category}
