% Generated by roxygen2: do not edit by hand
% Please edit documentation in multigmmsamedistribu/multigmmsamedistribu.R
\name{multigmmsamedistribu}
\alias{multigmmsamedistribu}
\title{Fit Multi 2-Component Gaussian Mixture Model in same distribution with a Fixed Proportion}
\usage{
multigmmsamedistribu(x, lower = 0.1)
}
\arguments{
\item{x}{a list of numeric vector}

\item{lower}{the lower bound of sigma}
}
\description{
Fit Multi 2-Component Gaussian Mixture Model in same distribution with a Fixed Proportion
}
\examples{
set.seed(0)
x1=c(rnorm(150,mean=0),rnorm(50,mean=10))
x2=c(rnorm(150,mean=20),rnorm(50,mean=40))
x3=c(rnorm(150,mean=30),rnorm(50,mean=60))
x=list(x1,x2,x3)
multigmmsamedistribu(x)
}
\keyword{distribution}
