% Generated by roxygen2: do not edit by hand
% Please edit documentation in multigmmsamedistribulik/multigmmsamedistribulik.R
\name{multigmmsamedistribulik}
\alias{multigmmsamedistribulik}
\title{The sum of Log-Likelihoods of 1D Multi Same Distribution Gaussian Mixture Model}
\usage{
multigmmsamedistribulik(x)
}
\arguments{
\item{x}{a list of numeric vectors. par contains lambda, sigma, and the parameters for each GMM, delta_mu and mean_mu. For example, supposing there are two GMM, the par would be (lambda, sigma ,mean_mu1, mean_mu2, delta_mu1, delta_mu2).}
}
\description{
The sum of Log-Likelihoods of 1D Multi Same Distribution Gaussian Mixture Model
}
\examples{
set.seed(0)
x1=c(
 rnorm(100, mean=0)
 , rnorm(100, mean=1)
 )
x=list(x1)
multigmmsamedistribulik(x)(c(0.5, 1, 0.5, 1))
}
\keyword{distribution}
